managedMap has some additional refactoring
