--- demodisplay/demomap.html
@@ -1,62 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Leaflet Map</title>
-  <link rel="stylesheet" href="https://unpkg.com/leaflet/dist/leaflet.css" />
-  <style>
-    #mapnew { height: 400px; }
- .label-style {
-  background-color: transparent; /* Make the background transparent */
-  border: none;                /* Remove the default border */
-  font-weight: bold;           /* Make the text bold (optional) */
-  color: black;    /* Make the text black */
-  opacity: 1.0;
-  box-shadow: none;
-  font-family:'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-  font-size: 1.5em;
- }
-
- .leaflet-tooltip-top:before,
-.leaflet-tooltip-bottom:before,
-.leaflet-tooltip-left:before,
-.leaflet-tooltip-right:before {
-  border: none !important; 
-}
-  </style>
-</head>
-<body>
-  <div id="mapnew"></div>
-
-  <script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js"></script>
-  <script>
-    // Initialize the map
-    const map = L.map('mapnew').setView([39.9526, -75.1652], 14); // Centered on Philadelphia
-    const testIcon = L.icon({iconUrl: '../mapgraphics/CompassW.svg', iconSize: [35, 35]});
-
-    // Add OpenStreetMap tiles
-    L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
-      attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
-    }).addTo(map);
-
-    // Define marker locations and labels
-    const locations = [
-      { name: "47", coords: [39.9547, -75.1443] }, // 1010 McKean St
-      { name: "45", coords: [39.9526, -75.1652] }, // City Hall
-      { name: "42", coords: [39.9653, -75.1800] }  // Art Museum
-    ];
-
-    // Add markers with labels
-    locations.forEach(location => {
-    L.marker(location.coords, {icon: testIcon})
-    .addTo(map) 
-   // .bindPopup(location.name)  // Keep the popup for clicking
-    .bindTooltip(location.name, {
-      permanent: true,         // Always show the label
-      direction: 'top',       // Position the label above the marker
-      opacity: 1.0,
-      className: 'label-style' // Add a CSS class for styling
-    });
-});
-  </script>
-</body>
-</html>
