--- demodisplay/horizontal.css
@@ -1,156 +1,0 @@
-/* Demo styles */
-
-.showMe {
-    /* Highlight the element's content area */
-    background-color: rgba(255, 81, 0, 0.3); /* Light yellow with transparency */
-    
-    /* Highlight the border */
-    outline: 3px solid rgba(3, 197, 251, 0.3); /* Blue with transparency */
-  
-   }
-
-
-/* Body Styles */
-  body {
-    font-family: "Helvetica Neue", Helvetica, sans-serif;
-    line-height: 1.6;
-    font-size: large;
-    color: #99cc99; /* Light text for dark background */
-    background-color: #222; /* Dark background */
-  }
-
-/* Link Styles */
-a {
-    color: #0096c7; /* Teal link color for dark mode */
-    text-decoration: none;
-    }
-    
-a:hover {
-    text-decoration: underline;
-}
-
-
-/* Info boxes */
-
-.infoBox {
-    display: grid;
-}
-
-.infoText {
-    display: inline-flex;
-    align-items:baseline;
-    justify-content:left;
-    overflow:visible;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    flex-grow: 0;
-}
-
-    .busNumber {
-        font-size: 4vh;
-        font-weight: bolder;
-        justify-content: center;
-        padding:0.5em;
-        align-items: center;
-        letter-spacing: 0.06em;
-    }
-
-    .streetPosition {
-        font-size: 1em;
-        font-weight: normal;
-        word-wrap: break-word;
-    }
-
-/* Multi-bus Display With Direction */
-
-.busSingleDirectionPosition {
-    display: grid;
-    max-width: 95%;
-    grid-template-rows: min-content min-content;
-    row-gap: 0;
-    gap: 0;
-    width: min-content;
-    word-wrap:break-word;
-    align-self: start;
-    justify-items: left;
-
-}
-
-.directionHeader {
-    display: flex;
-    font-size: 1.2em;
-    min-width: 0;
-    text-align: center;
-    font-family: Verdana, Geneva, Tahoma, sans-serif;
-    letter-spacing: 0.6ch;
-}
-
-.statusHolder {
-    display: flex;
-    flex-wrap: wrap;
-    flex: 1;
-    justify-content:first baseline;
-}
-
-.statusLine {
-    display: grid;
-    gap: 20px;
-    grid-template-columns: min-content min-content min-content;;
-    min-width: fit-content;
-    max-width: 90%;
-    grid-template-rows: min-content;
-}
-
-/* Encapsulate multiple direction statuses */
-
-
-.statusLineWithAlertMessage {
-    display: grid;
-    gap: 1em;
-    grid-template-columns: min-content;
-    grid-template-rows: min-content 1fr;   
-    min-height: fit-content;
-    min-width: fit-content;
-    max-width: 100%;
-    width: fit-content;
-}
-
-    .alertMessage {
-        font-family: "Atkinson Hyperlegible", Courier, 'Courier New', monospace;
-        font-size: 0.9em;
-        line-height: 1;
-        max-width: 100%;
-        box-sizing: border-box;
-        hyphens: auto;
-    }
-
-    .statusLineWithAlertMessage .statusLine {
-        display: flex;
-      }
-    
-/* Stack multiple status boxes */
-
-.statusLineWithMessageGroups {
-    display: flex;
-    gap: 2em 2em;
-    flex-wrap: wrap;
-    flex: 1 0 auto;
-}
-
-.statusLineWithMessageGroups .statusLineWithAlertMessage {
-    flex: 1 0 auto;
-    border: 1px dotted blue;
-    padding: 1em;
-    border-style: dashed;
-    border-spacing: 10px;
-} 
-
-/* Colorize based on seats */
-
-.yesSeats {
-    color: green;
-}
-
-.someSeats {
-    /* background-color: lightyellow; */
-}
