--- demodisplay/horizontaldemo.html
@@ -1,240 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Horizontal Test</title>
-  <link rel="stylesheet" href="../horizontal.css">
-  <link rel="preconnect" href="https://fonts.googleapis.com">
-  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-  <link href="https://fonts.googleapis.com/css2?family=Atkinson+Hyperlegible:ital,wght@0,400;0,700;1,400;1,700&display=swap" rel="stylesheet">
-</head>
-<body>
-
-<br>
-<div class="statusLine">
-  <div class="infoBox">
-    <div class="infoText busNumber showMe">
-        45
-    </div>
-  </div>
-  <div class="busSingleDirectionPosition showMe">
-    <div class="directionHeader"> North</div>
-    <div class="statusHolder">
-      <div class="infoText streetPosition showMe">
-        11th & Snyder
-      </div>
-      <div class="infoText streetPosition showMe">
-        11th & Moore
-      </div>
-      <div class="infoText streetPosition showMe">
-        11th & Dickinson
-      </div>
-  
-      
-    </div>
-  </div>
-  <div class="busSingleDirectionPosition showMe">
-    <div class="directionHeader"> South</div>
-    <div class="statusHolder">
-      <div class="infoText streetPosition showMe">
-        12th & Tasker
-      </div>
-      <div class="infoText streetPosition showMe">
-        12th & Moore
-      </div>
-      <div class="infoText streetPosition showMe">
-        12th & Snyder
-      </div>  
-      
-    </div>
-  </div>
-</div>
-
-<br>
-<div class="statusLineWithAlertMessage">
-  <div class="statusLine">
-    <div class="infoBox">
-      <div class="infoText busNumber">
-          45
-      </div>
-    </div>
-    <div class="busSingleDirectionPosition">
-      <div class="directionHeader"> North</div>
-      <div class="statusHolder">
-        <div class="infoText streetPosition">
-          11th & Snyder
-        </div>
-        <div class="infoText streetPosition">
-          11th & Moore
-        </div>
-        <div class="infoText streetPosition">
-          11th & Dickinson
-        </div>
-    
-        
-      </div>
-    </div>
-    <div class="busSingleDirectionPosition">
-      <div class="directionHeader"> South</div>
-      <div class="statusHolder">
-        <div class="infoText streetPosition">
-          12th & Tasker
-        </div>
-        <div class="infoText streetPosition">
-          12th & Moore
-        </div>
-        <div class="infoText streetPosition">
-          12th & Snyder
-        </div>  
-        
-      </div>
-    </div>
-  </div>
-  <div class="alertMessage">
-    <span>
-      Trouble brewing! This is sample alert text. I'm not sure what the problem is with this line, but there's a problem.
-    </span>
-  </div>
-</div>
-
-<br>
-
-<div class="statusLineWithMessageGroups">
-  <div class="statusLineWithAlertMessage">
-    <div class="statusLine">
-      <div class="infoBox">
-        <div class="infoText busNumber">
-            45
-        </div>
-      </div>
-      <div class="busSingleDirectionPosition">
-        <div class="directionHeader"> North</div>
-        <div class="statusHolder">
-          <div class="infoText streetPosition">
-            11th & Snyder
-          </div>
-          <div class="infoText streetPosition">
-            11th & Moore
-          </div>
-          <div class="infoText streetPosition">
-            11th & Dickinson
-          </div>
-      
-          
-        </div>
-      </div>
-      <div class="busSingleDirectionPosition">
-        <div class="directionHeader"> South</div>
-        <div class="statusHolder">
-          <div class="infoText streetPosition">
-            12th & Tasker
-          </div>
-          <div class="infoText streetPosition">
-            12th & Moore
-          </div>
-          <div class="infoText streetPosition">
-            12th & Snyder
-          </div>  
-          
-        </div>
-      </div>
-    </div>
-    <div class="alertMessage">
-      <span>
-        Trouble brewing! This is sample alert text. I'm not sure what the problem is with this line, but there's a problem.
-      </span>
-    </div>
-  </div>
-  <div class="statusLineWithAlertMessage">
-    <div class="statusLine">
-      <div class="infoBox">
-        <div class="infoText busNumber">
-            45
-        </div>
-      </div>
-      <div class="busSingleDirectionPosition">
-        <div class="directionHeader"> North</div>
-        <div class="statusHolder">
-          <div class="infoText streetPosition">
-            11th & Snyder
-          </div>
-          <div class="infoText streetPosition">
-            11th & Moore
-          </div>
-          <div class="infoText streetPosition">
-            11th & Dickinson
-          </div>
-      
-          
-        </div>
-      </div>
-      <div class="busSingleDirectionPosition">
-        <div class="directionHeader"> South</div>
-        <div class="statusHolder">
-          <div class="infoText streetPosition">
-            12th & Tasker
-          </div>
-          <div class="infoText streetPosition">
-            12th & Moore
-          </div>
-          <div class="infoText streetPosition">
-            12th & Snyder
-          </div>  
-          
-        </div>
-      </div>
-    </div>
-    <div class="alertMessage">
-      <span>
-        Trouble brewing! This is sample alert text. I'm not sure what the problem is with this line, but there's a problem.
-      </span>
-    </div>
-  </div>
-  <div class="statusLineWithAlertMessage">
-    <div class="statusLine">
-      <div class="infoBox">
-        <div class="infoText busNumber">
-            45
-        </div>
-      </div>
-      <div class="busSingleDirectionPosition">
-        <div class="directionHeader"> North</div>
-        <div class="statusHolder">
-          <div class="infoText streetPosition">
-            11th & Snyder
-          </div>
-          <div class="infoText streetPosition">
-            11th & Moore
-          </div>
-          <div class="infoText streetPosition">
-            11th & Dickinson
-          </div>
-      
-          
-        </div>
-      </div>
-      <div class="busSingleDirectionPosition">
-        <div class="directionHeader"> South</div>
-        <div class="statusHolder">
-          <div class="infoText streetPosition">
-            12th & Tasker
-          </div>
-          <div class="infoText streetPosition">
-            12th & Moore
-          </div>
-          <div class="infoText streetPosition">
-            12th & Snyder
-          </div>  
-          
-        </div>
-      </div>
-    </div>
-    <div class="alertMessage">
-      <span>
-        Trouble brewing! This is sample alert text. I'm not sure what the problem is with this line, but there's a problem.
-      </span>
-    </div>
-  </div>
-</div>
-
-</body>
-</html>
