--- demodisplay/inputcontrol/index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <script src="./demoinput.js" type="module"></script>
-    <head>
-
-    </head>
-    <body>
-        <div id="putThingsHere"> 
-            <form id="routeSubmissionForm">
-                <fieldset style="width:300px;">
-                    <label for="busRoute">Enter bus route:</label>
-                    <input type="search" id="busRouteInput" name="busRoute" 
-                    required minlength="1" maxlength="3" size="10" autofocus="true" 
-                    autocapitalize="characters" autocomplete="false" style="width: 44px; height: 44px;"/>
-                    <button type="submit" id="submitButton" style="width: 60px; height: 44px;">Submit</button>
-                    
-                    <div id="checkboxes">
-                        Lorem ipsum.
-                    </div>                        
-                </fieldset>            
-            </form>
-            
-        </div>    
-    </body>
-</html>
