--- demodisplay/leafletdemo/callservice.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Leaflet Managed Map Demo</title>
-  <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.4/dist/leaflet.css" />
-  <link rel="stylesheet" href="./leafletneeds.css" />
-  </style>
-</head>
-<body>
-  Howdy doody.
-  <div class="managed-map" id="mapdemo"></div>
-  <script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js"></script>
-  <script src="./managedmapdemo.js" type="module"></script>
-</body>
-</html>
