--- demodisplay/picklingdemo/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <script src="./picklingtest.js" type="module"></script>
-    <head>
-
-    </head>
-    <body>
-        <div id="Whatever">
-
-        </div>
-        <div id="Whatever2" style="word-wrap: break-word;">
-            
-        </div>
-    </body>
-</html>
