--- demodisplay/tableindex.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<title id="welcome-message">AutoBus</title>
-<link rel="stylesheet" href="horizontal.css">
-<!--- Below for favicon, design by me, files courtesy realfavicongenerator.net -->
-<link rel="apple-touch-icon" sizes="180x180" ref="favicon/apple-touch-icon.png">
-<link rel="icon" type="image/png" sizes="32x32" href="favicon/favicon-32x32.png">
-<link rel="icon" type="image/png" sizes="16x16" href="favicon/favicon-16x16.png">
-<link rel="manifest" href="favicon/site.webmanifest">
-<link rel="mask-icon" href="favicon/safari-pinned-tab.svg" color="#5bbad5">
-<meta name="msapplication-TileColor" content="#da532c">
-<meta name="theme-color" content="#ffffff">
-<meta name="viewport" content="width=device-width, initial-scale=0.7">
-<!--- End favicon -->
-</head>
-<body>
-    <h1>AutoBus</h1>
-    <div class="statusLineWithMessageGroups" id="demoArea">
-        \/ Put stuff here
-    </div>
-
-    <script src="visualscript.js" type="module"></script>
-</body>
-</html>
