+++ index.html
@@ -1,22 +1,29 @@
 <!DOCTYPE html>
 <html>
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
 <head>
-<title id="welcome-message">AutoBus</title>
-<link rel="stylesheet" href="style.css">
-<!--- Below for favicon, design by me, files courtesy realfavicongenerator.net -->
-<link rel="apple-touch-icon" sizes="180x180" ref="favicon/apple-touch-icon.png">
-<link rel="icon" type="image/png" sizes="32x32" href="favicon/favicon-32x32.png">
-<link rel="icon" type="image/png" sizes="16x16" href="favicon/favicon-16x16.png">
-<link rel="manifest" href="favicon/site.webmanifest">
-<link rel="mask-icon" href="favicon/safari-pinned-tab.svg" color="#5bbad5">
-<meta name="msapplication-TileColor" content="#da532c">
-<meta name="theme-color" content="#ffffff">
-<!--- End favicon -->
+  <title>Live Pushpin Painting Functionality</title>
+  <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.4/dist/leaflet.css" />
+  <link rel="stylesheet" href="./leafletneeds.css" />
+  <link rel="stylesheet" href="./visualstyling.css" />
+  </style>
 </head>
 <body>
-    <h1>AutoBus</h1>
-    <pre id="change-text">This is a simple static web app.</pre>
-
-    <script src="script.js" type="module"></script>
+  Pushpin painting live demo.
+  <div class="managed-map" id="simplemap"></div>
+  <div class="inputText" id="inputArea">
+    <form id="routeSubmissionForm">
+      <fieldset>
+        <label for="busRoute">Enter bus route:</label>
+        <input type="search" id="busRouteInput" name="busRoute" 
+        required minlength="1" maxlength="3" size="10" autofocus="true" 
+        autocapitalize="characters" autocomplete="false" style="width: 44px; height: 44px;"/>
+        <button type="submit" id="submitButton" style="width: 60px; height: 44px;">Submit</button>
+        <div class="checkbox-group" id="checkboxes"></div>
+      </fieldset>
+    </form>
+  </div>
+  <script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js"></script>
+  <script src="./visualdisplay.js" type="module"></script>
 </body>
 </html>
