--- service/tabledisplay/tabledisplayconstants.js
@@ -1,24 +1,0 @@
-const INFO_BOX = "infoBox";
-const INFO_TEXT = "infoText";
-
-const ROUTE_GROUP = "routeInfoDisplays";
-const STATUS_LINE_MESSAGE_GROUPS = "statusLineWithMessageGroups";
-    const SINGLE_ROUTE_STATUS = "statusLineWithAlertMessage";
-        const VEHICLE_POSITIONS = "statusLine";
-            const BUS_NUMBER = "busNumber";
-            const INDIVIDUAL_DIRECTION = "busSingleDirectionPosition";
-                const DIRECTION_HEADER= "directionHeader";
-
-                const STATUS_HOLDER = "statusHolder";
-                    const STREET_POSITION = "streetPosition";
-
-        const ALERT_MESSAGE = "alertMessage";
-
-
-const YES_SEATS_CSS = "yesSeats";
-const SOME_SEATS_CSS = "someSeats";
-
-export { INFO_BOX, INFO_TEXT, ROUTE_GROUP, SINGLE_ROUTE_STATUS, VEHICLE_POSITIONS, BUS_NUMBER,
-    INDIVIDUAL_DIRECTION, DIRECTION_HEADER, STATUS_HOLDER, STREET_POSITION, ALERT_MESSAGE,
-    STATUS_LINE_MESSAGE_GROUPS, SOME_SEATS_CSS, YES_SEATS_CSS
-}
