+++ visualdisplay.js
